Split key management test from kms test
Extract the key management related tests from the kms test file and put them in a new file

test: passed the python test
